+++ src/components/App/App.css
@@ -1,0 +1,6 @@
+.main {
+  padding: 0 20px;
+  font-family: "inter-400-regular", arial, sans-serif;
+  font-weight: 700;
+  font-size: 20px;
+}
