+++ src/components/App/App.js
@@ -1,0 +1,13 @@
+import "./App.css";
+import "normalize.css/normalize.css";
+import Header from "../../components/Header/Header";
+
+function App() {
+  return (
+    <div>
+      <Header />
+    </div>
+  );
+}
+
+export default App;
